Skip running PHPUnit with lowest dependencies on PHP 8.4
Many libraries, such as Symfony, don't understand semantic versioning
well, and tend to release software that is incompatible with the next
major version of a dependency (in this case `php`, which does not
follow SemVer).

This kind of botched architecture by Symfony leads to dependencies having
broken behavior in oldstable releases.

This is effectively a choice by the upstream maintainer (Symfony) to
leave this sort of problem to the consumer (in this case, us), who then
has to restrict ranges more aggressively.

In this case, the failures are due to transitive dependencies leading to
symfony, so introducing a `conflict` worsens the situation for our own
consumers/